Move platform detection into core extension provisioning
