+++ internal/command/extensions/sentry/create.go
@@ -2,7 +2,6 @@ package sentry_ext
 
 import (
 	"context"
-	"path/filepath"
 
 	"github.com/spf13/cobra"
 	"github.com/superfly/flyctl/gql"
@@ -11,7 +10,6 @@ import (
 	extensions_core "github.com/superfly/flyctl/internal/command/extensions/core"
 	"github.com/superfly/flyctl/internal/command/secrets"
 	"github.com/superfly/flyctl/internal/flag"
-	"github.com/superfly/flyctl/scanner"
 )
 
 func create() (cmd *cobra.Command) {
@@ -32,26 +30,7 @@ func create() (cmd *cobra.Command) {
 func runSentryCreate(ctx context.Context) (err error) {
 	appName := appconfig.NameFromContext(ctx)
 
-	absDir, err := filepath.Abs(".")
-
-	if err != nil {
-		return err
-	}
-
-	srcInfo, err := scanner.Scan(absDir, &scanner.ScannerConfig{})
-
-	if err != nil {
-		return err
-	}
-
-	options := gql.AddOnOptions{}
-
-	if srcInfo != nil && PlatformMap[srcInfo.Family] != "" {
-		options["platform"] = PlatformMap[srcInfo.Family]
-	}
-
 	sentryOptions := extensions_core.MonitoringExtensionDefaults
-	sentryOptions.Options = options
 
 	extension, err := extensions_core.ProvisionExtension(ctx, appName, sentryOptions)
 	secrets.DeploySecrets(ctx, gql.ToAppCompact(extension.App), false, false)
